all long debug economy a bit more: no less death. todo:  - add "savings", and invest in industry only from that savings & invest less if there are a big % of pop inside. Becaus right now, pop are making themselves starve when changing their job. And toomany money variations  - understand why wood_goods aren't created.

+++ src/remi/ssp_basegame/economy/FisheryIndustry.java
@@ -27,7 +27,7 @@ public class FisheryIndustry extends Industry {
 	
 	
 	@Override
-	public int produce(ProvinceIndustry indus, Collection<Pop> pops, int durationInDay) {
+	public long produce(ProvinceIndustry indus, Collection<Pop> pops, int durationInDay) {
 		Province prv = indus.getProvince();
 		if(!prv.isCoastal()) return 0;
 		
@@ -41,16 +41,16 @@ public class FisheryIndustry extends Industry {
 		}
 		
 		//TODO: use boats, and other goods to increase production
-		int production = 0;
+		long production = 0;
 		for(Pop pop : pops){
 			//1 tile : 1,66 men, 2 tiles: 2 mens, 3t: 2.33, 4t: 2.66, 5t: 3  (max) 6t: 3.33 mens
-			production += (int)( (2+seaTiles/2) * pop.getNbMensEmployed().getInt(indus) * 1);
+			production += (int)( (2+seaTiles/2) * pop.getNbMensEmployed().getLong(indus) * 1);
 		}
 		
 		//TODO set a rendement décroissant pour modeliser la surpeche
 		
 		// produce
-		int intproduction = myBasicNeeds.useGoodsAndTools(indus, (int)production * durationInDay, durationInDay);
+		long intproduction = myBasicNeeds.useGoodsAndTools(indus, (long)production * durationInDay, durationInDay);
 		super.sellProductToMarket(prv, intproduction, durationInDay);
 
 		return intproduction;
